Split extension methods into separate files
