+++ src/Generator.Rewrite/Method/MethodConstructor.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 using Mono.Cecil.Rocks;
+using OpenTK.Rewrite.Extensions;
 
 namespace OpenTK.Rewrite.Method
 {
