+++ src/Generator.Rewrite/Method/MethodConstructor_Debug.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 using Mono.Cecil.Rocks;
+using OpenTK.Rewrite.Extensions;
 
 namespace OpenTK.Rewrite.Method
 {
