+++ src/Generator.Rewrite/Method/MethodConstructor_Parameters.cs
@@ -5,6 +5,7 @@ using System.Runtime.InteropServices;
 using Mono.Cecil;
 using Mono.Cecil.Cil;
 using Mono.Cecil.Rocks;
+using OpenTK.Rewrite.Extensions;
 
 namespace OpenTK.Rewrite.Method
 {
